+++ test/optimizer/level-2/merge-media-queries-test.js
@@ -37,24 +37,24 @@ vows.describe('merge media queries')
         '@media (min-width:1024px){body{width:100%}}@media screen{a{color:red}div{display:block}}'
       ],
       'same two with breaking properties in between': [
-        '@media screen{a{color:red}}.one{color:#00f;display:inline}@media screen{div{display:block}}',
-        '@media screen{a{color:red}}.one{color:#00f;display:inline}@media screen{div{display:block}}'
+        '@media screen{.block-1{color:red}}.block-2{color:#00f;display:inline}@media screen{.block-3{display:block}}',
+        '@media screen{.block-1{color:red}}.block-2{color:#00f;display:inline}@media screen{.block-3{display:block}}'
       ],
       'same two with breaking @media in between': [
-        '@media screen{a{color:red}}@media (min-width:1024px){.one{color:#00f;display:inline}}@media screen{div{display:block}}',
-        '@media screen{a{color:red}}@media (min-width:1024px){.one{color:#00f;display:inline}}@media screen{div{display:block}}'
+        '@media screen{.block-1{color:red}}@media (min-width:1024px){.block-2{color:#00f;display:inline}}@media screen{.block-3{display:block}}',
+        '@media screen{.block-1{color:red}}@media (min-width:1024px){.block-2{color:#00f;display:inline}}@media screen{.block-3{display:block}}'
       ],
       'same two with breaking nested @media in between': [
-        '@media screen{a{color:red}}@media (min-width:1024px){@media screen{.one{color:#00f;display:inline}}}@media screen{div{display:block}}',
-        '@media screen{a{color:red}}@media (min-width:1024px){@media screen{.one{color:#00f;display:inline}}}@media screen{div{display:block}}'
+        '@media screen{.block-1{color:red}}@media (min-width:1024px){@media screen{.block-2{color:#00f;display:inline}}}@media screen{.block-3{display:block}}',
+        '@media screen{.block-1{color:red}}@media (min-width:1024px){@media screen{.block-2{color:#00f;display:inline}}}@media screen{.block-3{display:block}}'
       ],
       'intermixed': [
         '@media screen{a{color:red}}@media (min-width:1024px){p{width:100%}}@media screen{div{display:block}}@media (min-width:1024px){body{height:100%}}',
         '@media screen{a{color:red}div{display:block}}@media (min-width:1024px){p{width:100%}body{height:100%}}'
       ],
       'same two with overriding shorthand in between': [
-        '@media screen{a{font-size:10px}}@media (min-width:1024px){.one{font:13px Helvetica;display:inline}}@media screen{div{display:block}}',
-        '@media screen{a{font-size:10px}}@media (min-width:1024px){.one{font:13px Helvetica;display:inline}}@media screen{div{display:block}}'
+        '@media screen{.block-1{font-size:10px}}@media (min-width:1024px){.block-2{font:13px Helvetica;display:inline}}@media screen{.block-3{display:block}}',
+        '@media screen{.block-1{font-size:10px}}@media (min-width:1024px){.block-2{font:13px Helvetica;display:inline}}@media screen{.block-3{display:block}}'
       ],
       'same two with different component property in between': [
         '@media screen{a{font-size:10px}}@media (min-width:1024px){.one{font-weight:700}}@media screen{div{display:block}}',
@@ -81,8 +81,8 @@ vows.describe('merge media queries')
         '@media (max-width:768px){.one{padding-right:0;margin:0}}.one{padding:10px}'
       ],
       'backward of two with overriding shorthand in between': [
-        '@media screen{a{font-size:10px}}@media (min-width:1024px){.one{font:13px Helvetica}}@media screen{div{display:block}}',
-        '@media screen{a{font-size:10px}div{display:block}}@media (min-width:1024px){.one{font:13px Helvetica}}'
+        '@media screen{.block-1{font-size:10px}}@media (min-width:1024px){.block-1{font:13px Helvetica}}@media screen{.block-3{display:block}}',
+        '@media screen{.block-1{font-size:10px}.block-3{display:block}}@media (min-width:1024px){.block-1{font:13px Helvetica}}'
       ],
       'with at sign in rule name': [
         '@media screen{.one{margin-left:-16px}}.two{margin-left:0}.three{width:50%!important}.four{width:25%!important}@media screen{.three\\\@m{width:50%!important}.four\\\@m{width:25%!important}}',